move filter logic to own file
